+++ src/dbconform/conform.py
@@ -401,6 +401,7 @@ class DbConform:
             model_schema = ModelSchema.from_models(
                 models,
                 target_schema=self._target_schema,
+                schema_normalizer=dialect,
             )
             db_schema = DatabaseSchema.from_connection(connection, self._target_schema)
             differ = SchemaDiffer()
@@ -577,6 +578,7 @@ class AsyncDbConform:
             model_schema = ModelSchema.from_models(
                 models,
                 target_schema=self._target_schema,
+                schema_normalizer=dialect,
             )
             db_schema = await DatabaseSchema.from_connection_async(connection, self._target_schema)
             differ = SchemaDiffer()